feat: atualizando o Mardown da atividade da Aula6 do Sábado Letivo

+++ Aula6 + Segundo Sabado Letivo/CRUD_COM_ERROS/excluir.php
@@ -2,7 +2,13 @@
 // Exclusão com risco de SQL Injection e sem confirmação
 include("conexao.php");
 
-$id = $_GET["id"];
+
+$id = isset($_GET["id"]) ? $_GET["id"] : null;
+
+if ($id === null) {
+    die("ID não especificado.");
+}
+
 $sql = "DELETE FROM usuarios WHERE id = $id";
 mysqli_query($conn, $sql);
